u32 should be unsigned int

+++ SourceCode/Libraries/LibC/include/stdint.h
@@ -5,8 +5,8 @@ typedef unsigned char uint8_t;
 typedef unsigned char u8;
 typedef unsigned short uint16_t;
 typedef unsigned short u6;
-typedef unsigned long uint32_t;
-typedef unsigned long u32;
+typedef unsigned int uint32_t;
+typedef unsigned int u32;
 typedef unsigned long long uint64_t;
 typedef unsigned long long u64;
 
@@ -14,8 +14,8 @@ typedef signed char int8_t;
 typedef signed char s8;
 typedef signed short int16_t;
 typedef signed short s16;
-typedef signed long int32_t;
-typedef signed long s32;
+typedef signed int int32_t;
+typedef signed int s32;
 typedef signed long long int64_t;
 typedef signed long long s64;
